feat(frontend): fix the reset issue as per review comment
