+++ public/scripts/filters.js
@@ -1,3 +1,5 @@
+import { invokeMapResultsFormFilters } from './location.js';
+
 const filtersInstance = 'search_results';
 
 const searchResultSortFormId = 'sort_results';
@@ -326,6 +328,74 @@ const attachSearchResultsSortChangeListener = () => {
   }
 };
 
+/**
+ * Set the values for map result form compared with its value with search result form.
+ * @param {string} mapResultInstance
+ * @param {string} searchResultInstance
+ */
+const setMapFilterFormValues = (mapResultInstance, searchResultInstance) => {
+  const searchResutlForm = document.getElementById(`filters-${searchResultInstance}`);
+  const mapResultForm = document.getElementById(`filters-${mapResultInstance}`);
+
+  Array.from(searchResutlForm).forEach((element) => {
+    const elementName = element.name;
+    const elementValue = element.value;
+    const mapResultFormElements = mapResultForm.querySelector('[name="' + elementName + '"]');
+    const checkboxValue = elementValue === 'all' ? elementName : elementValue;
+    const mapResultCheckedElements = document.getElementById(`filters-${checkboxValue}-map_results`);
+
+    if (element.type === 'checkbox' && mapResultCheckedElements) {
+      mapResultCheckedElements.checked = element.checked;
+    }
+
+    if (element.type === 'text' && mapResultFormElements) {
+      mapResultFormElements.value = elementValue;
+    }
+  });
+  // manually checked the checkbox value for retired archived from search results form to map results form
+  document.getElementById(`filters-retired-archived-${mapResultInstance}`).checked = document.getElementById(
+    `filters-retired-archived-${searchResultInstance}`,
+  ).checked;
+};
+
+/**
+ * Add the subnit form event listner for map results page
+ * @param {string} instance
+ */
+const addMapFilterFormSubmitListener = (instance) => {
+  const form = document.getElementById(`filters-${instance}`);
+
+  form.addEventListener('submit', (e) => {
+    e.preventDefault();
+    const data = getValidatedFormData(form, instance);
+    invokeMapResultsFormFilters(true, data);
+  });
+};
+
+/**
+ * Add the reset form event listner for map results page
+ * @param {string} instance
+ */
+const addMapFilterFormResetListener = (instance) => {
+  const formSubmit = document.getElementById(`filters-${instance}`);
+
+  formSubmit.addEventListener('reset', (e) => {
+    e.preventDefault();
+    // reset all the map form fields
+    formSubmit.reset();
+    const checkboxes = document.querySelectorAll('.govuk-checkboxes__input');
+    checkboxes.forEach((checkbox) => {
+      checkbox.checked = false;
+    });
+    document.getElementById('filters-date-before-map_results').value = '';
+    document.getElementById('filters-date-after-map_results').value = '';
+    document.getElementById('filters-licence-map_results').value = '';
+    document.getElementById('filters-keywords-map_results').value = '';
+    invokeMapResultsFormFilters(true);
+    document.getElementById(`filters-retired-archived-map_results`).checked = false;
+  });
+};
+
 document.addEventListener('DOMContentLoaded', () => {
   attachSearchResultsSortChangeListener();
 
@@ -343,4 +413,7 @@ export {
   filtersInstance,
   addScopeChangeListener,
   getValidatedFormData,
+  setMapFilterFormValues,
+  addMapFilterFormResetListener,
+  addMapFilterFormSubmitListener,
 };
